Fix some bad variable name, commenting, fix inconsistency between float and Numeric, fix on datetime

+++ backend/app/models/airport.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Boolean, Index, Numeric, String
+from sqlalchemy import Boolean, Float, Index, String
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.db.database import Base
@@ -11,8 +11,8 @@ class Airport(Base):
     name: Mapped[str] = mapped_column(String(255), nullable=False)
     city: Mapped[str] = mapped_column(String(255), nullable=False)
     country: Mapped[str] = mapped_column(String(100), nullable=False)
-    latitude: Mapped[float] = mapped_column(Numeric(10, 6), nullable=False)
-    longitude: Mapped[float] = mapped_column(Numeric(10, 6), nullable=False)
+    latitude: Mapped[float] = mapped_column(Float, nullable=False)
+    longitude: Mapped[float] = mapped_column(Float, nullable=False)
     is_active: Mapped[bool] = mapped_column(Boolean, default=True)
 
     # idx_airports_coords as index to have faster query on radius
